Integrated SourceForge changes and successfully recompiled all programs.  Addressed the TransimsNet problem in Linux and made the following additional changed:
SysLib
The default constructor for Static_Service was changed from 100 to 0 to avoid redefining the traveler scaling factor when classes are derived from the service.  This affected a data class in AdjustPlans

PlanSelect 4.0.8
The travel time calculation with turning delays was made more consistent with the Router.

SubareaPlans 4.0.14
The travel time calculation with turning delays was made more consistent with the Router.

PlanSum 4.0.16
The option to save turning movement data to the link delay file was added.

AdjustPlans 4.0.5
The updated Static Service was included in the travel data class to fix a problem with traveler scaling factors that were not 100.