SysLib
MAX_DWELL was added to the vehicle type file, data, and service structures.

The ArcView file class was modified to automatically delete spatial index files (*.sbn, *.sbx) when creating a new shapefile.

Microsimulator 4.0.19
Enhanced the traveler output service to track transit vehicles as well as passengers.  Corrections were made to the transit processing rate calculations and max dwell was added to force vehicles to leave a station even if additional passengers are waiting to board.

TripSum 4.0.0
Converted to Version 4.0 format.

PlanSelect 4.0.9
The lane use file is now disabled when network processing is not needed.