April 20, 2007 Release Bug Fix
PopSyn 4.0.1
The PUMS vehicle or age fields were forced into the output file even if they are not included in the output field lists.  The field numbers are reset to the internal database values.  A bug was fixed in the output household field logic.

Router 4.0.19
A correction was made to the activity file processing to exit partitioned processing after the specified number of households was processed rather than the number of travelers.

Microsimulator 4.0.24
A bug in the problem file class caused several fields to be excluded from the output problem file.