ArcProblem/Release and ProblemSum/Release were put under management control
Test was removed from management control

The SysLib.vcproj was modified to use "$(ProjectDir)Include" in the debug version to facilitate use of SysLib.sln as a stand alone project.

PopSyn 4.0.5
Logic was added to avoid problems caused by not including a person number field in the PUMS population file.