SysLib Trip_Sum data and array classes were added to the Demand/Data directory.
ActGen 4.0.9
The Length_Data class was removed from the Data subdirectory and replaced with the Trip_Sum_Data class from the SysLib.

Microsimulator 4.0.28
The cycle failure calculation was changed to only count vehicles that are stopped on the link at the start of the green phase to avoid inaccuracies on long approach links.

ArcPlan 4.0.10
A bug was fixed for applications that generated an accessibility file, but not travel time contours.

PlanSum 4.0.26
The TRAVEL_SUMMARY_REPORT was modified to use the Trip_Sum_Data class and report minimum, maximum, average, and standard deviation for the travel time, trip length, travel speed and number of turns by time of day.  The option to output a NEW_PLAN_FILE with travel times updated based on the input link delay information was added.  A bug was fixed in generating a new turning movement file.  The time increment logic was fixed for situations where an input link delay file is provided that contains different time increments than specified by the summary.

LocationData 4.0.0
New program to add new fields to an activity location file and apply a user script to allocate data from a series of zone files to the activity locations within each zone.

PlanCompare 4.0.7
The TRAVEL_TIME_CHANGE_DISTRIBUTION report was added to summarize the distribution of travel time differences by minutes and percent of travelers.  Additional logic was added to ensure a proper match by trip leg.  This avoids problems created when one plan file includes access legs, but the other file does not.  NEW_DISTRIBUTION_FILE has been added to dump the travel time differences by hour of the day to a file.

PlanSelect 4.0.13
Additional error checks were added to avoid null pointers when skimming travel paths that are incompatible with the directional links (i.e., the network changed after the paths were built).

LinkSum 4.0.12
Turn summaries were added to the network performance report.

ArcDelay 4.0.2
The logic was updated to include the newer data fields in the performance-based link delay file in the data processing routine.  The density averaging calculation was fixed and the turn delay calculation was improved.

LinkDelay 4.0.6
Data structures and processing methods were added to combine all data fields in a performance-based link delay file.  Protection against potential integer overflow problems was included.