SysLib The logic to repair a definition file that does not match the corresponding dBase file was corrected.
A Smooth_Shape function was added to the utility library to address illogical shape points that result in bandwidth and offset distortions.

IntControl 4.0.9
A bug was fixed for situations where a new detector or signal coordinator key is not provided.  A bug was fixed for updates without time ranges.  The update logic was changed to use the maximum of the minimum and maximum green times in calculating the cycle length.  Updates for dual ring signals were corrected to properly separate rings and barriers in the calculations. Demand for right on red phases were excluded from the timing updates.

Microsimulator 4.0.29
The transit vehicle ID calculation was corrected to use the highest vehicle ID rather than the number of vehicle records.  The transit vehicle logic was improved in situations where a stop lane is time of day restricted due to parking or other use codes. 

ArcPlan 4.0.11
MAXIMUM_SHAPE_ANGLE and MINIMUM_SHAPE_LENGTH keys and the smooth shape function were added to refine the outside shape in bandwidth processing.

PlanCompare 4.0.8
PERCENT_PATH_DIFFERENCE, MINIMUM_PATH_DIFFERENCE, MAXIMUM_PATH_DIFFERENCE keys were added to constrain the selection to travelers based on significant path changes in addition to travel time changes.

ArcDelay 4.0.2
The logic was updated to include the newer data fields in the performance-based link delay file in the data processing routine.  The density averaging calculation was fixed and the turn delay calculation was improved.

GISNet 4.0.4
MAXIMUM_SHAPE_ANGLE and MINIMUM_SHAPE_LENGTH keys and the smooth shape function were added to refine link shapes and intersection approaches.