TRANSIMS Version 4.1 All changes are based on TRANSIMS 4.0 July 24, 2007 Software
SysLib
Execution
Message_Service was changed to a new default version and copyright string.

Demand/Data
The Plan Data class was implemented with Household-Person keys and Data_Array legs.  This class is the same as the Version 5 Plan Data class with one important exception.  The name of the Plan_Data class was renamed to Plan_Trip to avoid naming conflicts with the Plan_Data structure used by the Version 3/4 Plan File class.

Programs
Router 4.1.0
The Plan Data class was added for plan processing.  The Read_Trip, Read_Activity, and Build_Paths methods were modified to insert trip requests into the Plan Array and then call the new method Plan_Build to construct the paths and call the new method Write_Plan to save the plans to the output plan file in Version 3/4 format.  The Save_Plan method was modified to store path segments into the Plan_Legs of the Plan Data class.  (Note: the Write_Plan method is not complete at this time).