bootstrap and configure.ac.in: The bootstrap script creates the Makefile.am, Makefile.in and configure.ac files needed to build TRANSIMS on UNIX platforms.
Transims 4.1 8-1.doc:
UNIX make does not deal well with filenames that contain spaces so this file was renamed to Transims41/Transims- 4.1.8-1.doc.

SysLib/Lib:
This directory was remove because it only contained files generated by the build process.

svn:ignore property:
The svn:ignore property was added to each of the source directories so that svn would ignore files generated by the bootstrapping script or as part of the build process.