January 2008 Release SysLib: The help flag was expanded to include a list of report options.
A bug was fixed in the Rider Data class and logic was added to the ridership service to interpolate the departure time for each stop based on actual and schedule times.

The Max_Dwell and Min_Dwell methods within the Vehicle_Type data class were rounded.

The Output_Ridership class was corrected to interpret the schedule and time methods as Microsimulator time steps.

The FIELD_BUFFER set at 256 could not be used in dBase/ArcView files.  The string fields for notes, vehicle use restrictions, and street names were set back to 255 for newly created files.

LineSum 4.0.1
The subarea selection polygon logic was implemented.  The reports were modified slightly to accommodate outputting a subset of stops on a transit line.  The time period method options were expanded to include both actual (RUN_START, RUN_MIDPOINT, RUN_END) and schedule (SCHEDULE_START, SCHEDULE_MIDPOINT, and SCHEDULE_END) times.

ActGen 4.0.12
Several data-related error messages were converted to warning messages.  A new report called LOCATION_CHOICE_DETAILS_* was included to dump out detailed status messages for selected households during the location choice process.

SideFriction 4.0.0
New program to generate Lane_Use records for short lane blockages associated with double parking, bus stops, and other side friction events.

Reschedule 4.0.1
The option to input a Ridership file and/or a Vehicle Type file was include.  This enables the schedule adjustments to record the actual stop departure times from the Microsimulator or to calculate the stop dwell time based on the boarding and alighting processing rates.  SELECT_TRANSIT_MODES and SELECT_TRANSIT_ROUTES were added to constrain the routes that are rescheduled.

Microsimulator 4.0.37
The transit process rates were adjusted to work correctly with sub-second time steps.  The Output_Ridership class was corrected to interpret the schedule and time methods as time steps.

PlanSum 4.0.32
A bug was fixed in the time period calculation for transit ridership summaries.  The Vehicle Type file was added to calculate the stop dwell time based on boarding and alighting processing rates.  Compatibility problems between the plans and the network now use the Warning interface.  The average and maximum density fields in the output link delay file were re-calculated using 24 hours when the summary time increment key is zero.

TPPlusNet 4.0.2
Logic was added to detect and automatically convert the speeds in the speed-capacity lookup table from miles per hour to meters per second.

ArcNet 4.0.11
DRAW_AB_DIRECTION key was added to draw one-way links with lanes only in the B to A direction with shape points in the A to B direction.  

GISNet 4.0.5
GIS_NODE_FILE key was added to initialize the node numbers for link centerline conversions.