Router:
The methods that compute routes were moved from the Router class to a new
Plan_Builder class.  Movement of this code was in preparation for the
parallelization of the Router.  In the future, independent trips may be routed
simultaneously by utilizing multiple Plan_Builder objects along with multiple
threads and/or processes.

-----

SysLib:

Some of the data and constants needed by the routing routines existed in
classes defined in SysLib.  Unfortunately, the classes containing those data
and constants ultimately inherited the Static_Service and Execution_Service
classes.  Since only one instance of the Static_Service and Execution_Service
classes may exist within a program, the data and constants had to be extracted
into standalone classes.  These new classes were then inherited by the classes
from while they were extracted.  This allowed the Router and all other existing
TRANSIMS programs to continue using the classes in SysLib as they had done
before, and allowed multiple instances of the Plan_Builder to use the new
classes access to the data and constants.

The 'const' modifier was introduced to the parameters and return values of some
methods as appropriate.  These modifications were made largely to detect the
misuse of references to constants returned by a method but also detect other
programming errors within a method.  Some methods were not modified, despite
being obvious candidates, because the modifications produced errors in far too
many locations.  This issue should be addressed more thoroughly in the future.