SysLib MAX_WARNING_MESSAGES and MAX_WARNING_EXIT_FLAG keys were added to the Execution_Service, Report_Service, and Message_Service interfaces.  Default behavior will terminate the execution after 100,000 warning messages.
CONICAL and EXPONENTIAL volume-delay functions were added.

PlanSelect 4.0.19
A bug was fixed in setting the flags for SELECT_OD_COORDINATES and EXCLUDE_OD_COORDINATES

ActGen 4.0.15
An error message was removed for situations where the vehicle file includes vehicles for households that are not in the household file.  TRAVEL_TIME_FIELD_MODE_x key was added to enable the specification of a skim field for use in the travel time calculations for each mode.

PlanSum 4.0.34
Conical and exponential volume-delay functions were added.

Router 4.0.28
Conical and exponential volume-delay functions were added.

ProblemSelect 4.0.3
The input household list file was made optional.