SysLib Several memory delete calls were replaced with free calls.
ProblemSum 4.0.1
A bug was fixed in printing the summary increment value.  The increment value was also changed from seconds to minutes to be consistent with other summary programs.  A bug was fixed in the selection logic for situations were the problem density report, select links, and select polygons were not specified.

ArcProblem 4.0.1
The time increment value was changed from seconds to minutes to be consistent with other summary programs.  The logic was changed to require an activity location file when the problem file does not include a link field.  A bug was fixed in the activity location offset calculation.

PlanSum 4.0.35
The skim and trip table summary logic was modified to avoid printing warning messages when processing incomplete transit trips.  A bug was fixed in the path based selection logic for node-based paths with zero nodes.