ExportTransit 64bit compile commands were updated.
SysLib
A bug was fixed in the Db_Base logic for converting string fields into integers and doubles.  The Date_Time class was also enhanced to better handle time strings that don’t contain time related characters.

Router 4.0.32
The restricted link logic was expanded to permit the path to include a continuous sequence of restricted links starting or ended at a restricted origin and/or destination