TPPlusRoute 4.0.2 Logic was added to combine the headways and travel times for routes that are merged and have service in the same time period.  A bug was fixed when PERIOD_TRAVEL_TIMES is false.
ArcNet 4.0.15
Logic was added to copy the name and notes fields from the transit route file to the arcview shape file.

TransitNet 4.0.13
HOURLY and DAILY parking cost fields were added to the input park-&-ride node file.

Router 4.0.39
MAX_KISS_RIDE_DROPOFF_WALK and KISS_RIDE_STOP_TYPES keys were added to permit the user to control the maximum distance between a parking lot and a transit stop that will be considered for kiss-&-ride trips, and control the types of transit stops were kiss-&-ride trips are permitted.  A bug was fixed in the fare class distribution control interface.  A bug was fixed for park-&-ride/kiss-&-ride trips were the activity location and parking lot record ordering are different for a given location.

ArcNet 4.0.17
The link retrieval logic for Route Nodes files with negative node numbers was fixed and the message when a link between two nodes is not found was changed from an error message to a warning message.