SysLib The length units in the Smooth_Shape function were corrected.
TransimsNet 4.0.9
A bug was fixed in reading the signal warrant keys.  Added logic to avoid creating stop signs on walk and rail links.  The signal warrant logic was improved for dividing arterials and intersections with only one entry option.  External activity locations and parking lots are no longer added to external links that do not permit autos and do permit rail.

TransitNet 4.0.14
A bug was fixed in the XY coordinate calculated for the stop activity locations.  Logic was added to search for a cross street that permits walking when the transit link does not permit walking.  Warning messages were added when activity locations are added to links that do not permit walking.  Protection was added for negative stop offsets caused by stops on very short links.  Logic was added to generate external stops and activity locations on external access links.