SysLib Logic was added to the network Read_Zone method to print the highest zone number in the report.
The problem code number was added to the problem summary report.

PlanSum 4.0.37
A bug was fixed in the time periods for transit link group reports.

ConvertTrips 4.0.16
Location weight field checks and automatic corrections are no longer performed. The network zone file was included to optionally check the zone data against zone numbers in the zone file.  The network link file was included to optionally check that the activity location is on an auto accessible link before generating a warning about missing parking lots.