SysLib The network read services for link, lane-use, parking, turn, toll, and transit stop files were modified to process blank use codes more effectively.  The node-based turn prohibition key was fixed.
ActGen 4.0.19
ACTIVITY_FILE key was added to read an existing activity file and update the activities and trip data for households in the household list.  All other households are copied to the output activity file.

Microsimulator 4.0.45
Added percent calculations to the number of vehicle and transit trips started and completed.

Validate 4.0.14
Additional Facility Type labels were added to the Facility Type report.

PlanSelect 4.0.21
The network is no longer disabled if only time ratios are selected.

PlanSum 4.0.38
A bug was fixed in the number of time periods allocated for turning movement summaries.

TransimsNet 4.0.10
FACILITY_TYPE_ACCESS_FLAGS key was added to enable the user to override the default facility types where parking lots and activity locations are generated.  The logic for utilizing turn prohibition records was corrected.  NEW_TURN_PROHIBITION_TABLE key was added to output the input turn prohibitions to the new network directory and file format.