ConvertTrips 4.0.17 Logic was added to check for drive modes before requiring an activity location to have a parking lot attached.
TransimsNet 4.0.10
The logic for transit-related external links was modified to add the walk use code if walk is not already permitted.

IntControl 4.0.11
PRINT_MERGE_WARNINGS key was added to enable or disable the warning messages about merge or diverge movements at intersections.

ArcPlan 4.0.18
SELECT_RANDOM_PERCENTAGE and RANDOM_NUMBER_SEED keys were added to select a random percentage of the travelers selected based on the result of the other selection criteria.