June 2008 Release
Microsimulator 4.0.47
The logic was changed to avoid problems caused when transit vehicles fail to load onto the network.  The look-ahead lane changing logic was modified to include the lane change penalty for the initial lane change.

TransitNet 4.0.15
The exit status for lane connectivity warning messages was changed to an error rather than a warning.

TransimsNet 4.0.11
Logic was added to properly label complex diverge configurations and the lane connectivity for multi-lane left side diverge movements was improved.

LinkSum 4.0.19
A bug was fixed in the top 100 volume report.

LocationData 4.0.5
BOUNDARY_POLYGON_# keys were added to associate activity locations within various polygons so that the conversion script can set variables or make calculations.

Router 4.0.43
A bug was fixed in setting the traveler ID for the first activity in an activity file tour.