SysLib The control list class was enhanced to facilitate reading multiple control files.
RunSetup 4.0.5
The batch and control file naming convention was modified to include run, alternative, year, and model replacement keys in the file prefix if provided.  A run name is no longer required and period or underscore separators are now more effectively removed from control keys when run, alternative, year, or model replacement keys are not provided.  SETUP_CONTROL_FILE_# and SETUP_PARAMETER_# keys were added to enable RunSetup to process a list of RunSetup control files to create a single batch file for a multi-step process.