SysLib Extended file class was modified to include enhanced partitioning logic.
TripSum 4.0.6
The household list file was implemented with the enhanced extended file logic to permit partitioned household lists to be used to partition an input activity or trip file.  Partitioned household list files can also be used to select a subset of records from a set of partitioned input activity or trip files.