SysLib Protection was added to the trip length data array to avoid problems caused by negative travel times or distances.
A Location_Flag method was added to the snapshot file class to include X_COORD, Y_COORD, Z_COORD, and BEARING fields in Version 4.0 file formats.

The OUTPUT_SNAPSHOT_LOCATION_FLAG key was added to the snapshot generation class used in the Microsimulator to include location data (x, y, z, and bearing) in the Version 4.0 file formats.  The azimuth field is not calculated for the VERSION3 format. 

TripSum 4.0.8
TIME_OF_DAY_FORMAT key was added specify the time format for the input trip or activity file.  The extra protection in the trip length data array was included to avoid problems caused by negative travel times or distances.

Microsimulator 4.0.49
The OUTPUT_SNAPSHOT_LOCATION_FLAG key was added to include location data (x, y, z, and bearing) in the Version 4.0 file formats.