SysLib SQLITE3 database software was added to SysLib and edited so it compiles in 32bit and 64bit versions without warning messages.
Router 4.0.49
VEHICLE_TYPE_FILE key was added to enable Version 4.0 vehicle type use codes to be used in the control path building permissions.  The ROUTE_SELECTED_MODES key was modified to accept range codes.

TripSum 4.0.9
SELECT_ORIGIN_POLYGON and SELECT_DESTINATION_POLYGON keys were added to select trips based on the coordinates of the origin and/or destination activity locations.  The selection criteria (household list, modes, purposes, origins, and destinations) now control the records output to the new trip or activity file as well as the data used in the summary reports.  The select modes key was modified to accept range codes.

ArcSnapshot 4.0.8
NEW_SNAPSHOT_FILE, NEW_SNAPSHOT_FORMAT, SNAPSHOT_FORMAT, and OCCUPANCY_FORMAT keys were added.  The new snapshot file can be generated with or without ArcView snapshot files.  The X and Y coordinates and Bearing fields included in the output snapshot file are based on the link shape and lane offset of the vehicle.