PlanSum 4.0.40 Traveler IDs were added to transit-related warning messages.
SubareaPlans 4.0.25
A minor incompatibility between SubareaNet and SubareaPlans in processing transit paths at subarea boundaries was corrected.  Logic was added to properly handle the trip and leg number for travelers with multiple input trips.  Refinements were included to avoid loading transit trips on boundary links that are leaving the subarea when the boarding stop is outside of the subarea boundary. The fare zone for the external stops was set based on the fare zone assigned to the closest internal stop. A bug was fixed for trips that only utilize stops on the new transit route that is generated when a regional route enters and exits the subarea twice. EXTERNAL_OFFSET_LENGTH key was added to enable the user to control the distance away from the external node where external station activity locations, parking lots, and transit stops are inserted.   Internal-External summary statistics for transit legs were added.

SubareaNet 4.0.7
The fare zone for the external stops was set based on the fare zone assigned to the closest internal stop.  EXTERNAL_OFFSET_LENGTH key was added to enable the user to control the distance away from the external node where external station activity locations, parking lots, and transit stops are inserted. 

Router 4.0.50
One meter was added to walk between activity locations on the same link to minimize round-off error and encourage walk paths to use the network links rather than direct movements between activity locations.