SysLib The process link processing method was modified to keep transit stop connections when value checks are not requested.
PlanTrips 4.0.6
The program was updated to process transit trips as well as drive trips. It also processes multi-leg plans and outputs the plan records for each leg of trips with travel schedule adjustments.

ArcPlan 4.0.19
The selection logic was enhanced to permit a non-partitioned household list file to be used with partitioned plan file.