The AECOM branch is ready for October 2008 Release preparation.
SysLib
The pointer array classes were restructured and renamed.  The new data pointer array was integrated into the time equivalence class methods. 

The DONEX enumeration value was added to the exit_stat options to override the automatic program destructor call.  This was used in PopSyn to avoid potential destructor errors that could not be tracked down at this time.

The SQLITE3 SQL logic was modified to use single quotes rather than double quotes to identify string fields.  This eliminated the need to add and subtract the double quotes from the strings on input and output.

ActGen 4.0.25
NEAREST_NEIGHBOR_FACTOR and NEAREST_NEIGHBOR_FIELD keys were added to the zone skim group to enable the program to calculate intrazonal travel values based on a factor of the values from the nearest neighboring zone.

PopSyn 4.0.7
The supporting data structures were converted to pointer arrays in order to minimize problems with the program destructors.  Unfortunately these changes did not completely resolve the problem, so the DONEX exit_stat option was used to skip the destructor call at this point in time.