SysLib The zone file interface was modified to include the option for minimum and maximum coordinate ranges.
ActGen 4.0.27
The zone processing was modified to read the optional minimum and maximum coordinate ranges from the zone file.  The minimum and maximum factor calculations were changed to represent the ratio of the minimum and maximum straight-line distance to the four corners of the zone coordinate range or the zone centroid to the straight-line distance between the origin and destination zone centroids.