SysLib An error message was added when attempting to open an SQLITE3 database as a text file.
Vehicle file processing was changed to report a warning message and skip the record rather than an error message when a vehicle parking lot was not found. 

TransitNet 4.0.16
The logic was changed to avoid inserting a stop on a short link when the previous link has a facility type (freeway or other) warning, but the stop was inserted anyway.

SubareaPlans 4.0.26
Error messages about vehicle location problems were changed to warning messages.