SysLib The data reading methods were modified to move the required file checks into the read methods and out of the record processing methods.
A new problem type related to household person matches was added to the problem service interface.

Several data classes and arrays previously included in the ActGen data directory were moved to the SysLib directories for use by other problems.

ArcGen 4.0.30
A bug was fixed in the maximum adjustment factors for the return trip for anchor activities.  The household, person, and vehicle file processing was modified to reduce memory requirements.  The input processing performance for zone skim files was improved by as much as a factor of 20 by pre-allocating the maximum memory. The SKIM_MEMORY_FACTOR_x key was added to refine the maximum memory allocation for zone skim files.  NEW_ACTIVITY_PATTERN_FILE and NEW_ACTIVITY_PATTERN_FORMAT keys were added to generate an activity pattern file for input to the new LocationChoice program or manipulation by the new ActivityPattern program.  Household person match problems were included in the household match report and output file and added to the problem summary and problem file.

ActivityPattern 4.0.0
The household matching to survey households, persons, and activities from ActGen was extracted into a new program.  This program outputs an activity pattern file.

LocationChoice 4.0.0
The location choice and scheduling components of ActGen was extracted into a new program.  This program reads an activity pattern file and creates an activity file.

ModeChoice 4.0.3
The input processing performance for zone skim files was improved by as much as a factor of 20 by pre-allocating the maximum memory. The SKIM_MEMORY_FACTOR_x key was added to refine the maximum memory allocation for zone skim files.