SysLib The connection types for merge and diverge were subdivided into left and right merges and splits.
TransimsNet 4.0.16
The program was restructured to work in update mode.  UPDATE_NODE_RANGE and UPDATE_LINK_RANGE keys were added to enable the user to select specific nodes and links to re-calculate.  Data for the records not updated are copied from input network files.  The NET_ACTIVITY_LOCATION_TABLE, NET_PARKING_TABLE, NET_PROCESS_LINK_TABLE, NET_POCKET_LANE_TABLE, NET_UNSIGNALISED_NODE_TABLE, and NET_SIGNALIZED_NODE_TABLE keys were activated to provide the input data for update applications.  Input records for the updated nodes or links are deleted by the copy process and the updated data are added to the end of the output files. The merge and diverged connection types were subdivided into left and right merges and splits.  The lane connectivity logic was enhanced to accommodate merges and diverges better. A bug was fixed in the bearing calculations for links with lanes only in the BA direction.

ArcDelay 4.0.15
SYSTEM_EVENT_FILE key was added in order to calculate green, yellow and red time splits for the ArcView turn file.  A bug was fixed for applications where a turning movement summary is the only output shape file.

ActGen 4.0.31
Improvements were made for tours that do not return home at the end of the day.

LocationChoice 4.0.1
Improvements were made for tours that do not return home at the end of the day.