ActGen 4.0.32 The zone number, activity location, and purpose of the two anchor points for the location choice rubber banding were added to the tour information record passed to the location choice script.  For the option where a zone skim travel time field is used to define the travel time between two activity locations, the skim time is adjusted based on the relative distances between the two activity locations and the two zone centroids.  For intrazonal trips, the intrazonal distance is based on half of the maximum and minimum X and Y coordinates assigned to the zone.
LocationChoice 4.0.2
The zone number, activity location, and purpose of the two anchor points for the location choice rubber banding were added to the tour information record passed to the location choice script.  For the option where a zone skim travel time field is used to define the travel time between two activity locations, the skim time is adjusted based on the relative distances between the two activity locations and the two zone centroids. For intrazonal trips, the intrazonal distance is based on half of the maximum and minimum X and Y coordinates assigned to the zone.

TransimsNet 4.0.17
The update logic was corrected to read existing pocket lane information when only the nodes are updated.