TRANSIMS Version 4.0 -- November 2008 Release From AECOM Branch 748
SysLib
An error message was added when attempting to open an SQLITE3 database as a text file.

Vehicle file processing was changed to report a warning message and skip the record rather than an error message when a vehicle parking lot was not found. 

The data reading methods were modified to move the required file checks into the read methods and out of the record processing methods.

A new problem type related to household person matches was added to the problem service interface.

Several data classes and arrays previously included in the ActGen data directory were moved to the SysLib directories for use by other problems. 

The connection types for merge and diverge were subdivided into left and right merges and splits.

A warning message was added for transit routes with fare zones outside the range specified in the transit fare file.

TransitNet 4.0.16
The logic was changed to avoid inserting a stop on a short link when the previous link has a facility type (freeway or other) warning, but the stop was inserted anyway.

SubareaPlans 4.0.26
Error messages about vehicle location problems were changed to warning messages.

ArcGen 4.0.32
The dump report now includes the minimum and maximum adjustment factors for each zone interchange.  A bug was fixed in the maximum adjustment factors for the return trip for anchor activities.  The household, person, and vehicle file processing was modified to reduce memory requirements.  The input processing performance for zone skim files was improved by as much as a factor of 20 by pre-allocating the maximum memory. The SKIM_MEMORY_FACTOR_x key was added to refine the maximum memory allocation for zone skim files.  NEW_ACTIVITY_PATTERN_FILE and NEW_ACTIVITY_PATTERN_FORMAT keys were added to generate an activity pattern file for input to the new LocationChoice program or manipulation by the new ActivityPattern program.  Household person match problems were included in the household match report and output file and added to the problem summary and problem file.  Improvements were made for tours that do not return home at the end of the day.  The zone number, activity location, and purpose of the two anchor points for the location choice rubber banding were added to the tour information record passed to the location choice script.  For the option where a zone skim travel time field is used to define the travel time between two activity locations, the skim time is adjusted based on the relative distances between the two activity locations and the two zone centroids.  For intrazonal trips, the intrazonal distance is based on half of the maximum and minimum X and Y coordinates assigned to the zone.

ActivityPattern 4.0.0
The household matching to survey households, persons, and activities from ActGen was extracted into a new program.  This program outputs an activity pattern file.

LocationChoice 4.0.2
The location choice and scheduling components of ActGen was extracted into a new program.  This program reads an activity pattern file and creates an activity file.  Improvements were made for tours that do not return home at the end of the day. The zone number, activity location, and purpose of the two anchor points for the location choice rubber banding were added to the tour information record passed to the location choice script.  For the option where a zone skim travel time field is used to define the travel time between two activity locations, the skim time is adjusted based on the relative distances between the two activity locations and the two zone centroids. For intrazonal trips, the intrazonal distance is based on half of the maximum and minimum X and Y coordinates assigned to the zone.

CheckSurvey 4.0.3
Additional logic was added to check for and repair sequential activities that share the same location code.

ModeChoice 4.0.3
The input processing performance for zone skim files was improved by as much as a factor of 20 by pre-allocating the maximum memory. The SKIM_MEMORY_FACTOR_x key was added to refine the maximum memory allocation for zone skim files.  

LineSum 4.0.8
The selection set data class was changed to avoid multiple deletes of data pointers while closing the program.

TransimsNet 4.0.17
Improvements were made to the lane connectivity logic for multi-lane diverges on the left or right side.  The program was restructured to work in update mode.  UPDATE_NODE_RANGE and UPDATE_LINK_RANGE keys were added to enable the user to select specific nodes and links to re-calculate.  Data for the records not updated are copied from input network files.  The NET_ACTIVITY_LOCATION_TABLE, NET_PARKING_TABLE, NET_PROCESS_LINK_TABLE, NET_POCKET_LANE_TABLE, NET_UNSIGNALISED_NODE_TABLE, and NET_SIGNALIZED_NODE_TABLE keys were activated to provide the input data for update applications.  Input records for the updated nodes or links are deleted by the copy process and the updated data are added to the end of the output files. The merge and diverged connection types were subdivided into left and right merges and splits.  The lane connectivity logic was enhanced to accommodate merges and diverges better. A bug was fixed in the bearing calculations for links with lanes only in the BA direction.  The update logic was corrected to read existing pocket lane information when only the nodes are updated.

ArcDelay 4.0.15
SYSTEM_EVENT_FILE key was added in order to calculate green, yellow and red time splits for the ArcView turn file.  A bug was fixed for applications where a turning movement summary is the only output shape file.

ArcRider 4.0.5
The number of transit runs and load factor fields were added to the segment data generated for the ridership summary.  SET_WIDTH_USING_RUNS and SET_WIDTH_USING_LOAD_FACTOR keys were added to enable ridership bandwidth plots to be generated based on the number of transit runs on the link or the number of riders per run.

ArcDiff 4.0.5
The program was expanded to generate difference files for ArcView link direction and ridership files as well as link delay files.  New keys include ARCVIEW_LINK_DIR_FILE_1, ARCVIEW_LINK_DIR_FILE_2, ARCVIEW_RIDERSHIP_FILE_1, ARCVIEW_RIDERSHIP_FILE_2, ARCVIEW_LINK_DIR_FILE_1_*, ARCVIEW_LINK_DIR_FILE_2_*, ARCVIEW_RIDERSHIP_FILE_1_*,  and ARCVIEW_RIDERSHIP_FILE_2_*.