SysLib The data range class was modified to accept negative values (i.e., directional links) in the range list.
PlanSelect 4.0.23
SELECT_LINKS_# and SELECT_TRAVEL_MODES keys were added to select highway and transit trips based on links and/or mode codes.  The SELECT_NODES_# behavior was enhanced to mimic the link selection logic that enables transit passengers to be selected based on the driver paths of the transit vehicles.