The MPI parallelization environment has been changed so that a normal compilation produces binaries that only perform serially. There are new configurations for DebugMPI and ReleaseMPI that compile the TRANSIMS Microsimulator with MPI support as long as the Microsoft Cluster Computing SDK is installed under Visual Studio. The executables are place into the directories DebugMPI and BinMPI.
Also, a control key was added to the Microsimulator (USE_DYNAMIC_REROUTING) to allow turning on the dynamic rerouting feature in the Microsimulator. By default, dynamic rerouting is turned off and the Microsimulator behaves like it does in the standard TRANSIMS version.

MPI-enabled routing is not yet supported under Linux.