SysLib The related file logic was modified to only check for optional files when check data and renumbering flags are set.  It no longer requires that data records exist for these files.
The Num_Records method for SQLITE3 was modified to account for the actual number of records in the table rather than the maximum record number.

Relocate 4.0.1
New program to update activity locations and parking lots in trip, activity, and vehicle files that are removed or renumbers as a result of TransimsNet (or manual) network edits.  PLAN_FILE, PLAN_FORMAT, NEW_PLAN_FILE, and NEW_PLAN_FORMAT keys were added to implement plan file processing as well.  The logic was also modified to permit the processing of trip and activity files (along with plan and vehicle files) at the same time.